+++ src/app/(operation-check)/operation-check/image/image-download/_components/download-image/fetch-download-url/index.ts
@@ -1,5 +1,5 @@
 import { env } from "@/env.mjs";
-import { Failure, Result, Success } from "@/lib/result/result";
+import { Failure, Result, Success } from "@/lib/result";
 import { z } from "zod";
 
 const schema = z.object({
