+++ src/app/(operation-check)/operation-check/image/image-upload/_components/image-upload-form/upload-image/index.tsx
@@ -1,4 +1,4 @@
-import { Failure, Result, Success } from "@/lib/result/result";
+import { Failure, Result, Success } from "@/lib/result";
 import { upload } from "@/lib/storage/upload";
 import imageCompression from "browser-image-compression";
