+++ src/app/api/counter/route.ts
@@ -1,6 +1,6 @@
 import { getCurrentUser } from "@/lib/auth/get-current-user";
-import { db } from "@/lib/db/db";
-import { Failure, Result, Success } from "@/lib/result/result";
+import { db } from "@/lib/db";
+import { Failure, Result, Success } from "@/lib/result";
 import { z } from "zod";
 
 const patchSchema = z.object({ count: z.number() });
