+++ src/lib/result/index.ts
